+++ tests/test_package/test_log.py
@@ -9,6 +9,8 @@ import tidy3d as td
 from tidy3d.exceptions import Tidy3dError
 from tidy3d.log import DEFAULT_LEVEL, _get_level_int, set_logging_level
 
+from ..utils import assert_log_level, log_capture, AssertLogLevel
+
 
 def test_log():
     td.log.debug("debug test")
@@ -268,3 +270,29 @@ def test_log_suppression():
         assert td.log._counts is None
 
     td.config.log_suppression = True
+
+
+def test_assert_log_level(log_capture):
+    """Test features of the assert_log_level"""
+
+    # log was captured
+    with AssertLogLevel(log_capture, "WARNING", contains_str="ABC"):
+        td.log.warning("ABC")
+
+    # string was not matched (manually clear because not sure any other way using context manager)
+    td.log.warning("ABC")
+    with pytest.raises(Exception):
+        assert_log_level(log_capture, "WARNING", contains_str="DEF")
+    log_capture.clear()
+
+    # string was matched at the wrong level
+    td.log.info("ABC")
+    with pytest.raises(Exception):
+        assert_log_level(log_capture, "WARNING", contains_str="ABC")
+    log_capture.clear()
+
+    # log exceeds expected level
+    td.log.warning("ABC")
+    with pytest.raises(Exception):
+        assert_log_level(log_capture, "INFO")
+    log_capture.clear()