+++ tidy3d/components/medium.py
@@ -11,6 +11,7 @@ import numpy as np
 import xarray as xr
 
 from .base import Tidy3dBaseModel, cached_property
+from .base import skip_if_fields_missing
 from .grid.grid import Coords, Grid
 from .types import PoleAndResidue, Ax, FreqBound, TYPE_TAG_STR
 from .types import InterpMethod, Bound, ArrayComplex3D, ArrayFloat1D
@@ -546,6 +547,7 @@ class AbstractMedium(ABC, Tidy3dBaseModel):
     )
 
     @pd.validator("modulation_spec", always=True)
+    @skip_if_fields_missing(["nonlinear_spec"])
     def _validate_modulation_spec(cls, val, values):
         """Check compatibility with modulation_spec."""
         nonlinear_spec = values.get("nonlinear_spec")
@@ -1102,6 +1104,7 @@ class Medium(AbstractMedium):
     )
 
     @pd.validator("conductivity", always=True)
+    @skip_if_fields_missing(["allow_gain"])
     def _passivity_validation(cls, val, values):
         """Assert passive medium if `allow_gain` is False."""
         if not values.get("allow_gain") and val < 0:
@@ -1113,6 +1116,7 @@ class Medium(AbstractMedium):
         return val
 
     @pd.validator("permittivity", always=True)
+    @skip_if_fields_missing(["modulation_spec"])
     def _permittivity_modulation_validation(cls, val, values):
         """Assert modulated permittivity cannot be <= 0."""
         modulation = values.get("modulation_spec")
@@ -1127,6 +1131,7 @@ class Medium(AbstractMedium):
         return val
 
     @pd.validator("conductivity", always=True)
+    @skip_if_fields_missing(["modulation_spec", "allow_gain"])
     def _passivity_modulation_validation(cls, val, values):
         """Assert passive medium if `allow_gain` is False."""
         modulation = values.get("modulation_spec")
@@ -1239,14 +1244,12 @@ class CustomIsotropicMedium(AbstractCustomMedium, Medium):
         return val
 
     @pd.validator("conductivity", always=True)
+    @skip_if_fields_missing(["permittivity"])
     def _conductivity_real_and_correct_shape(cls, val, values):
         """Assert conductivity is real and of right shape."""
 
         if val is None:
             return val
-
-        if values.get("permittivity") is None:
-            raise ValidationError("'permittivity' failed validation.")
 
         if not CustomIsotropicMedium._validate_isreal_dataarray(val):
             raise SetupError("'conductivity' must be real.")
@@ -1256,6 +1259,7 @@ class CustomIsotropicMedium(AbstractCustomMedium, Medium):
         return val
 
     @pd.validator("conductivity", always=True)
+    @skip_if_fields_missing(["allow_gain"])
     def _passivity_validation(cls, val, values):
         """Assert passive medium if `allow_gain` is False."""
         if val is None:
@@ -1475,6 +1479,7 @@ class CustomMedium(AbstractCustomMedium):
         return val
 
     @pd.validator("eps_dataset", always=True)
+    @skip_if_fields_missing(["modulation_spec", "allow_gain"])
     def _eps_dataset_eps_inf_greater_no_less_than_one_sigma_positive(cls, val, values):
         """Assert any eps_inf must be >=1"""
         if val is None:
@@ -1523,6 +1528,7 @@ class CustomMedium(AbstractCustomMedium):
         return val
 
     @pd.validator("permittivity", always=True)
+    @skip_if_fields_missing(["modulation_spec"])
     def _eps_inf_greater_no_less_than_one(cls, val, values):
         """Assert any eps_inf must be >=1"""
         if val is None:
@@ -1546,14 +1552,12 @@ class CustomMedium(AbstractCustomMedium):
         return val
 
     @pd.validator("conductivity", always=True)
+    @skip_if_fields_missing(["permittivity", "allow_gain"])
     def _conductivity_non_negative_correct_shape(cls, val, values):
         """Assert conductivity>=0"""
 
         if val is None:
             return val
-
-        if values.get("permittivity") is None:
-            raise ValidationError("'permittivity' failed validation.")
 
         if not CustomMedium._validate_isreal_dataarray(val):
             raise SetupError("'conductivity' must be real.")
@@ -1572,6 +1576,7 @@ class CustomMedium(AbstractCustomMedium):
         return val
 
     @pd.validator("conductivity", always=True)
+    @skip_if_fields_missing(["eps_dataset", "modulation_spec", "allow_gain"])
     def _passivity_modulation_validation(cls, val, values):
         """Assert passive medium at any time during modulation if `allow_gain` is False."""
 
@@ -2021,6 +2026,7 @@ class DispersiveMedium(AbstractMedium, ABC):
         """Assert modulated permittivity cannot be <= 0 at any time."""
 
         @pd.validator("eps_inf", allow_reuse=True, always=True)
+        @skip_if_fields_missing(["modulation_spec"])
         def _validate_permittivity_modulation(cls, val, values):
             """Assert modulated permittivity cannot be <= 0."""
             modulation = values.get("modulation_spec")
@@ -2041,6 +2047,7 @@ class DispersiveMedium(AbstractMedium, ABC):
         """Assert passive medium at any time if not ``allow_gain``."""
 
         @pd.validator("modulation_spec", allow_reuse=True, always=True)
+        @skip_if_fields_missing(["allow_gain"])
         def _validate_conductivity_modulation(cls, val, values):
             """With conductivity modulation, the medium can exhibit gain during the cycle.
             So `allow_gain` must be True when the conductivity is modulated.
@@ -2633,10 +2640,9 @@ class CustomPoleResidue(CustomDispersiveMedium, PoleResidue):
         return val
 
     @pd.validator("poles", always=True)
+    @skip_if_fields_missing(["eps_inf"])
     def _poles_correct_shape(cls, val, values):
         """poles must have the same shape."""
-        if values.get("eps_inf") is None:
-            raise ValidationError("'eps_inf' failed validation.")
 
         expected_coords = values["eps_inf"].coords
         for coeffs in val:
@@ -2808,6 +2814,7 @@ class Sellmeier(DispersiveMedium):
     )
 
     @pd.validator("coeffs", always=True)
+    @skip_if_fields_missing(["allow_gain"])
     def _passivity_validation(cls, val, values):
         """Assert passive medium if `allow_gain` is False."""
         if values.get("allow_gain"):
@@ -2964,6 +2971,7 @@ class CustomSellmeier(CustomDispersiveMedium, Sellmeier):
         return val
 
     @pd.validator("coeffs", always=True)
+    @skip_if_fields_missing(["allow_gain"])
     def _passivity_validation(cls, val, values):
         """Assert passive medium if `allow_gain` is False."""
         if values.get("allow_gain"):
@@ -3133,6 +3141,7 @@ class Lorentz(DispersiveMedium):
         return val
 
     @pd.validator("coeffs", always=True)
+    @skip_if_fields_missing(["allow_gain"])
     def _passivity_validation(cls, val, values):
         """Assert passive medium if `allow_gain` is False."""
         if values.get("allow_gain"):
@@ -3321,11 +3330,9 @@ class CustomLorentz(CustomDispersiveMedium, Lorentz):
         return val
 
     @pd.validator("coeffs", always=True)
+    @skip_if_fields_missing(["eps_inf"])
     def _coeffs_correct_shape(cls, val, values):
         """coeffs must have consistent shape."""
-        if values.get("eps_inf") is None:
-            raise ValidationError("'eps_inf' failed validation.")
-
         expected_coords = values["eps_inf"].coords
         for de, f, delta in val:
             if (
@@ -3355,6 +3362,7 @@ class CustomLorentz(CustomDispersiveMedium, Lorentz):
         return val
 
     @pd.validator("coeffs", always=True)
+    @skip_if_fields_missing(["allow_gain"])
     def _passivity_validation(cls, val, values):
         """Assert passive medium if `allow_gain` is False."""
         allow_gain = values.get("allow_gain")
@@ -3580,11 +3588,9 @@ class CustomDrude(CustomDispersiveMedium, Drude):
         return val
 
     @pd.validator("coeffs", always=True)
+    @skip_if_fields_missing(["eps_inf"])
     def _coeffs_correct_shape_and_sign(cls, val, values):
         """coeffs must have consistent shape and sign."""
-        if values.get("eps_inf") is None:
-            raise ValidationError("'eps_inf' failed validation.")
-
         expected_coords = values["eps_inf"].coords
         for f, delta in val:
             if f.coords != expected_coords or delta.coords != expected_coords:
@@ -3697,6 +3703,7 @@ class Debye(DispersiveMedium):
     )
 
     @pd.validator("coeffs", always=True)
+    @skip_if_fields_missing(["allow_gain"])
     def _passivity_validation(cls, val, values):
         """Assert passive medium if `allow_gain` is False."""
         if values.get("allow_gain"):
@@ -3805,11 +3812,9 @@ class CustomDebye(CustomDispersiveMedium, Debye):
         return val
 
     @pd.validator("coeffs", always=True)
+    @skip_if_fields_missing(["eps_inf"])
     def _coeffs_correct_shape(cls, val, values):
         """coeffs must have consistent shape."""
-        if values.get("eps_inf") is None:
-            raise ValidationError("'eps_inf' failed validation.")
-
         expected_coords = values["eps_inf"].coords
         for de, tau in val:
             if de.coords != expected_coords or tau.coords != expected_coords:
@@ -3822,6 +3827,7 @@ class CustomDebye(CustomDispersiveMedium, Debye):
         return val
 
     @pd.validator("coeffs", always=True)
+    @skip_if_fields_missing(["allow_gain"])
     def _passivity_validation(cls, val, values):
         """Assert passive medium if `allow_gain` is False."""
         allow_gain = values.get("allow_gain")
@@ -4172,6 +4178,7 @@ class FullyAnisotropicMedium(AbstractMedium):
         return val
 
     @pd.validator("conductivity", always=True)
+    @skip_if_fields_missing(["permittivity"])
     def conductivity_commutes(cls, val, values):
         """Check that the symmetric part of conductivity tensor commutes with permittivity tensor
         (that is, simultaneously diagonalizable).
@@ -4189,6 +4196,7 @@ class FullyAnisotropicMedium(AbstractMedium):
         return val
 
     @pd.validator("conductivity", always=True)
+    @skip_if_fields_missing(["allow_gain"])
     def _passivity_validation(cls, val, values):
         """Assert passive medium if `allow_gain` is False."""
         if values.get("allow_gain"):